# enhance(scripts/diff-features): improve performance