#### Summary

Improves the performance of the `diff-features` script (used by the `release` script):

1. Only fetch the workflow run for the `head_sha` we're looking for.
2. Fetch changes asynchronously.

#### Test results and supporting details

Ran `time npm run release -- --simulate` before and after:

```console
# Before
8.60s user 3.49s system 9% cpu 2:08.19 total

# After
7.26s user 3.28s system 44% cpu 23.535 total
```

#### Related issues

<!-- 🔨 If applicable, use "Fixes #XYZ" -->

<!-- ✅ After submitting, review the results of the "Checks" tab! -->